fix: put some prompt instructions at the bottom
Instructions at the bottom of the prompt are always more effective. In
this case "Do not include quotation marks" was not always being followed
by Turbo. This commit reproduces the original python mobile's version
where this was put after the character defs, not before:

"The mood in the message should be ${mood}, and talk about ${topic}. It should be short, 2-3 lines. Do not include quotation marks."